// client/src/interfaces/IExpenses.tsx
export interface Expenses {
    propertyTaxes: number;
    insurance: number;
    repairsAndMaintenance: number;
    vacancy: number;
    capitalExpenditures: number;
    managementFees: number;
    electricity: number;
    gas: number;
    water: number;
    HOAfees: number;
    garbage: number;
    other: number;
}